Add 0.1mm solder mask expansion.
Should have always been there, but perhaps wasn’t included in the Notecarrier-F port because it gets overridden by a board wide setting anyway. But is required in the Notecarrier-A port so a good opportunity to fix it and use the part for both projects. Also added ref des text to the Fab layer for same purpose, but hidden by default.

Note the same two improvements will also be made to SOT666 and J-NANOSIM-SF72S006VBA, but they are rolled up with other changes to those footprints, so don't appear in this commit.