Add models, footprints and symbols from Notecarrier-F KiCad port rev B.
Includes all the new 3D models (see the Porting Notes in the Notecarrier-F KiCad project for details) as well as the alternate M.2 symbol “MDT420E01001_ALT”.

All the footprints are named with a “_Fv1.2” suffix because they likely overlap with existing footprints, and because this is a one off it’s not clear what the merge behaviour should be. Distinguishing by name will make it easy to merge and consolidate in the future.