Correct MDT420E01001 (M.2 connector) pin 38 name “NC38” —> “AUX5”.
In Notecarrier-Pi it is NC38.
In Notecarrier-F it is NC38, but the net name is N_AUX5.
In Notecarrier-A it is NC38, but the net name is AUX5.
But in Notecarrier-B it is AUX5 and the net name is AUX5.

And now in reference design “SingleDryCell” it is again “AUX5”. Hence, I am updating the library since it will be used for “SingleDryCell”, and looks very much like a correction. In keeping with the original projects, I have also updated Notecarrier-B, but none of the others.